+++ ogcapi-client/src/lib.rs
@@ -2,38 +2,38 @@
 //!
 //! # Example
 //!
-//! ```rust, ignore
+//! ```rust,ignore
+//! use futures_util::TryStreamExt;
 //! use ogcapi_client::Client;
-//! use ogcapi_types::common::Bbox;
-//! use ogcapi_types::stac::SearchParams;
 //!
-//! # fn main() {
-//! // Setup a client for a given STAC endpoint
-//! let endpoint = "https://data.geo.admin.ch/api/stac/v0.9/";
-//! let client = Client::new(endpoint).unwrap();
+//! #[tokio::main]
+//! async fn main() {
+//!     let client = Client::new("https://data.geo.admin.ch/api/stac/v0.9/").unwrap();
+//!     let root = client.root().await.unwrap();
+//!     println!("Root id: {}", root.id);
 //!
-//! // Fetch root catalog and print `id`
-//! let root = client.root().unwrap();
-//! println!("Root catalog id: {}", root.id);
-//!
-//! // Count catalogs
-//! let catalogs = client.catalogs().unwrap();
-//! println!("Found {} catalogs!", catalogs.count());
-//!
-//! // Search items
-//! let bbox = Bbox::from([7.4473, 46.9479, 7.4475, 46.9481]);
-//! let params = SearchParams::new()
-//!     .with_bbox(bbox)
-//!     .with_collections(["ch.swisstopo.swissalti3d"].as_slice());
-//! let items = client.search(params).unwrap();
-//! println!("Found {} items!", items.count());
-//! # }
+//!     let mut collections = client.collections().await.unwrap();
+//!     while let Some(c) = collections.try_next().await.unwrap() {
+//!         println!("{}", c.id);
+//!     }
+//! }
+//! ```
 
 mod client;
 mod error;
 
-#[cfg(feature = "processes")]
-mod processes;
+#[cfg(feature = "blocking")]
+#[cfg(not(target_arch = "wasm32"))]
+mod blocking;
 
-pub use client::Client;
+#[cfg(feature = "processes")]
+pub mod processes;
+
+pub(crate) static UA_STRING: &str = "OGCAPI-CLIENT";
+
+pub use client::{Client, CollectionsStream, ItemsStream};
 pub use error::Error;
+
+#[cfg(feature = "blocking")]
+#[cfg(not(target_arch = "wasm32"))]
+pub use blocking::BlockingClient;